Refactor - change int handling in fee policy config parsing Minor changes in fee policy watcher Signed-off-by: Trayan Manolov <trayan.manolov@limechain.tech>

+++ app/model/fee-policy/flat_fee_per_token_policy.go
@@ -49,7 +49,7 @@ func ParseNewFlatFeePerTokenPolicy(networks []uint64, parsingValue interface{}) 
 		}
 
 		if foundToken && foundValue {
-			feeValue, ok := value.(int64)
+			feeValue, ok := value.(int)
 
 			if !ok {
 				return nil, errors.New("value is not integer")
